backend-api/requirements.txt
asgiref==3.8.1
boto3==1.35.67
botocore==1.35.67
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
Django==5.0.6
django-cors-headers==4.4.0
django-storages==1.14.4
djangorestframework==3.15.1
facenet-pytorch==2.6.0
filelock==3.15.1
fonttools==4.53.0
fsspec==2024.6.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python-headless==4.10.0.82
packaging==24.1
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.4
scipy==1.13.1
six==1.16.0
sqlparse==0.5.0
sympy==1.12.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.1
